refactor: modularize Apps model and Apps controller for init code refactor
