+++ models.py
@@ -15,9 +15,9 @@ class Users_Habits(SQLModel, table=True):
     name: str
 
 
-class Apps(SQLModel, table=True):
-    id: int | None = Field(default=None, primary_key=True)
-    name: str
-    description: str
-    link: str
-    user_id: int
+# class Apps(SQLModel, table=True):
+#     id: int | None = Field(default=None, primary_key=True)
+#     name: str
+#     description: str
+#     link: str
+#     user_id: int
